requirements.txt
appdirs==1.4.0
cffi==1.2.1
characteristic==14.3.0
cryptography==1.0.1
cssselect==0.9.1
decorator==4.0.2
enum34==1.0.4
funcsigs==0.4
idna==2.0
ipaddress==1.0.14
lxml==3.4.4
matplotlib==1.4.3
mock==1.3.0
networkx==1.10
nose==1.3.7
numpy==1.9.2
pandas==0.16.2
pbr==1.8.0
Pillow==2.9.0
progressbar2==2.7.3
py==1.4.30
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
pyOpenSSL==0.15.1
pyparsing==2.0.3
pytest==2.7.2
python-dateutil==2.4.2
pytools==2015.1.3
pytz==2015.6
queuelib==1.4.2
requests==2.7.0
scikit-image==0.11.3
scipy==0.16.0
Scrapy==1.0.3
service-identity==14.0.0
six==1.9.0
Twisted==15.4.0
virtualenv==13.1.2
w3lib==1.12.0
zope.interface==4.1.2
